# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/my/loans": {
      "get": {
        "description": "Retrieves the authenticated member's loans with computed fields (canRenew, isOverdue, daysUntilDue). Supports filtering by status and sorting. Member only.",
        "operationId": "MyLoansController_findMyLoans",
        "parameters": [
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Items per page (default: 20, max: 100)",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (default: 1)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort direction (default: asc)",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Field to sort by (default: dueDate)",
            "schema": {
              "enum": [
                "dueDate",
                "borrowedAt",
                "createdAt"
              ],
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by loan status",
            "schema": {
              "enum": [
                "PENDING",
                "APPROVED",
                "REJECTED",
                "BORROWED",
                "RETURNED",
                "OVERDUE",
                "CANCELLED"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Member's loans retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "loans": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "memberId": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "bookCopyId": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "PENDING",
                              "APPROVED",
                              "REJECTED",
                              "BORROWED",
                              "RETURNED",
                              "OVERDUE",
                              "CANCELLED"
                            ]
                          },
                          "requestedAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "approvedAt": {
                            "type": "string",
                            "format": "date-time",
                            "nullable": true
                          },
                          "borrowedAt": {
                            "type": "string",
                            "format": "date-time",
                            "nullable": true
                          },
                          "dueDate": {
                            "type": "string",
                            "format": "date-time",
                            "nullable": true
                          },
                          "returnedAt": {
                            "type": "string",
                            "format": "date-time",
                            "nullable": true
                          },
                          "renewalCount": {
                            "type": "number",
                            "example": 0
                          },
                          "overdueFee": {
                            "type": "number",
                            "example": 0
                          },
                          "rejectionReason": {
                            "type": "string",
                            "nullable": true
                          },
                          "bookCopy": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string",
                                "format": "uuid"
                              },
                              "copyNumber": {
                                "type": "string"
                              },
                              "book": {
                                "type": "object",
                                "properties": {
                                  "id": {
                                    "type": "string",
                                    "format": "uuid"
                                  },
                                  "title": {
                                    "type": "string"
                                  },
                                  "subtitle": {
                                    "type": "string",
                                    "nullable": true
                                  },
                                  "isbn": {
                                    "type": "string"
                                  },
                                  "coverImageUrl": {
                                    "type": "string",
                                    "nullable": true
                                  },
                                  "authors": {
                                    "type": "array",
                                    "items": {
                                      "type": "object"
                                    }
                                  }
                                }
                              }
                            }
                          },
                          "canRenew": {
                            "type": "boolean",
                            "description": "Whether loan can be renewed"
                          },
                          "isOverdue": {
                            "type": "boolean",
                            "description": "Whether loan is overdue"
                          },
                          "daysUntilDue": {
                            "type": "number",
                            "description": "Days until due (negative if overdue)",
                            "nullable": true
                          }
                        }
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "number",
                          "example": 1
                        },
                        "pageSize": {
                          "type": "number",
                          "example": 20
                        },
                        "totalItems": {
                          "type": "number",
                          "example": 5
                        },
                        "totalPages": {
                          "type": "number",
                          "example": 1
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized (MEMBER only)"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get current member's loans",
        "tags": [
          "My Loans"
        ]
      }
    },
...
  },
  "info": {
    "title": "Library Management System API",
    "description": "RESTful API for a comprehensive Library Management System. This API provides endpoints for managing books, authors, categories, book copies, members, loans, and system settings.",
    "version": "1.0.0",
    "contact": {
      "name": "Library Management Team",
      "url": "https://github.com/library-management",
      "email": "support@library-management.com"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development Server"
    },
    {
      "url": "https://api.library-management.com",
      "description": "Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
...
  }
}